+++ src/components/image/index.ts
@@ -3,10 +3,12 @@ export type {
   ImageProps,
   ImageSlot,
   NebaImageFilter,
+  NebaImageFlip,
   NebaImageFrame,
   NebaImageFrameOptions,
   NebaImageFrameShape,
   NebaImageProtection,
+  NebaImageRotation,
   NebaImageWatermark,
   NebaImageWatermarkOptions
 } from './Image.js';